Use mappers to handle arbitrary content types
We're using "mappers" to transform a given item that could be an event,
an entry, or some other kind of content, into a standard form that is
then recognized by the summary-content-list-item template.

For example, we want datetimes for both entries and events. However,
events don't have specific times, only date ranges. An example date
range looks like "2016-05-08⋯15" for an event that took place between
May 8th and May 15th in 2016. So for entries, we map a `.date` field
that contains the item's canonical date, while for events we map a
`.date` field that is the start of that range.

It remains to be seen if one mapper will be needed for each kind of
<target-template, content-type> pair, or if this is all the specificity
we'll need.